+++ ebcpy/simulationapi/__init__.py
@@ -173,7 +173,7 @@ class SimulationAPI:
         self._progress_int = 0
         # Handle deprecation warning
         self._working_directory = None  # Define instance attribute
-        self.working_directory = working_directory
+        self.working_directory = Path(working_directory).absolute()
         save_logs = kwargs.get("save_logs", True)
         self.logger = setup_logger(
             working_directory=self.working_directory if save_logs else None,
@@ -608,7 +608,7 @@ class SimulationAPI:
         :param str title:
             Title of the study
         :param pathlib.Path path:
-            Where to store the .zip file. If not given, self.working_directory is used.
+            Where to store the .zip file. If not given, self.cd is used.
         :param list files:
             List of files to save along the standard ones.
             Examples would be plots, tables etc.
@@ -618,7 +618,7 @@ class SimulationAPI:
             specified to avoid input during execution.
         """
         if path is None:
-            path = self.working_directory
+            path = self.cd
         return save_reproduction_archive(
             title=title,
             path=path,